Reorganize workflow engine based files
