+++ poc-tests/modules/plugins/influxdb_reporter/.gitignore
