+++ poc-tests/modules/plugins/influxdb_reporter/setup.py
